+++ RELEASES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Version 0.9
+
+- now support `leafwing-input-manager` 0.15
+- all types provided by this library are now `Reflect`
+- removed `ToggleActions`: this functionality no longer makes sense with changes to how LWIM disables actions. Use run conditions directly on the new `AbilitySystem::TickCooldowns` system set
+
 ## Version 0.8
 
 - now supports Bevy 0.14